pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
  { name = "Charlie Bini", email = "5003326+cbini@users.noreply.github.com" },
]
license-expression = "MIT"
dependencies = [
  "requests>=2.27.1",
  "google-cloud-storage>=2.2.1",
  "python-dateutil>=2.8.2",
]
requires-python = ">=3.9"

[project.urls]
Homepage = ""

[project.optional-dependencies]
[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = []

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"

[tool.pdm.scripts]
_.env_file = ".env"
extract = { call = "deanslist_sync.extract:main" }
post_lock = { shell = "pdm export -f requirements > requirements.txt" }
